+++ src/main/java/com/aiqin/bms/scmp/api/supplier/domain/request/supplier/vo/ApplySupplyCompanyReqVO.java
@@ -103,11 +103,11 @@ public class ApplySupplyCompanyReqVO {
 
     @NotNull(message = "最低订货金额不能为空")
     @ApiModelProperty("最低订货金额")
-    private Long minOrderAmount;
+    private BigDecimal minOrderAmount;
 
     @NotNull(message = "最高订货金额不能为空")
     @ApiModelProperty("最高订货金额")
-    private Long maxOrderAmount;
+    private BigDecimal maxOrderAmount;
     
     @ApiModelProperty("发货信息")
     private List<ApplyDeliveryInfoReqVO> deliveryInfoList;
